Fix mistakes and remarks
