+++ src/components/burger-constructor/burger-constructor.module.css
@@ -16,7 +16,8 @@
 
 .container > div {
     margin-right: 20px;
-    margin-left: 30px;
+    /*margin-left: 30px;*/
+    width: 475px;
 }
 
 .list {
@@ -39,10 +40,16 @@
     width: 100%;
 }
 
+.item > div {
+    width: 475px;
+}
+
 .item svg {
     margin-right: 10px;
 
     cursor: pointer;
+    width: 24px;
+    height: 24px;
 }
 
 .order {
