+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -1,56 +1,21 @@
-import React, {useEffect, useState} from "react";
+import React from "react";
 
 import styles from './burger-ingredients.module.css';
 
 import {TabPanel} from "../tab-panel/tab-panel";
 import {IngredientsList} from "../ingredients-list/ingredients-list";
-import {Modal} from "../modal/modal";
-import {DataContext} from "../../utils/appContext";
 
-import {URL, ERROR_MESSAGE} from "../../utils/constants";
 
 export const BurgerIngredients = () => {
-  const [data, setData] = useState( []);
-  const [errorMessage, setErrorMessage] = useState(false);
-
-  useEffect(() => {
-    fetch(URL)
-      .then(res => {
-        if (res.status !== 200) {
-          return Promise.reject(new Error(res.statusText))
-        }
-        return Promise.resolve(res);
-      })
-      .then(res => res.json())
-      .then(result => setData(result.data))
-      .catch((err) => setErrorMessage(true));
-  }, []);
-
-  const closeModal = () => {
-    setErrorMessage(false);
-  }
-
   return (
-    <DataContext.Provider value={{data, setData}}>
-      <section className={styles.wrapper}>
-        <h2 className="text text_type_main-large">Соберите бургер</h2>
-        <TabPanel />
-        <div className={`${styles.list} custom-scroll`}>
-          <IngredientsList type="bun"/>
-          <IngredientsList type="sauce"/>
-          <IngredientsList type="main" />
-        </div>
-      </section>
-      {errorMessage &&
-        <Modal closeModal={closeModal}>
-          <p className="text text_type_main-large">
-            {ERROR_MESSAGE}
-          </p>
-        </Modal>
-      }
-    </DataContext.Provider>
+    <section className={styles.wrapper}>
+      <h2 className="text text_type_main-large">Соберите бургер</h2>
+      <TabPanel />
+      <div className={`${styles.list} custom-scroll`}>
+        <IngredientsList type="bun"/>
+        <IngredientsList type="sauce"/>
+        <IngredientsList type="main" />
+      </div>
+    </section>
   )
-}
-
-BurgerIngredients.propTypes = {
 }
