+++ src/utils/constants.js
@@ -1,7 +1,6 @@
-export const URL = 'https://norma.nomoreparties.space/api/ingredients';
+export const API_URL = 'https://norma.nomoreparties.space/api';
 export const ERROR_MESSAGE = 'Что-то пошло не так! Перезагрузите страницу';
 
-export const URL_ORDER ='https://norma.nomoreparties.space/api/orders';
 export const ERROR_MESSAGE_ORDER = 'К сожалению не удалось оформить заказ. Попробуйте чуть позже';
 export const EMPTY_ORDER = 'Пока в вашем бургере ничего нет. Обязательно добавьте побольше всего вкусного';
